Enhance SuperDrawer code.
