+++ src/utils/index.ts
@@ -26,5 +26,4 @@ export function toGroupState<T extends string>(values: T[], onChange: (value: T[
   };
 }
 
-export * from "./SuperDrawer";
 export * from "./useTestIds";
